requirements.txt
Django==5.2.3
djangorestframework==3.16.0
PyMySQL==1.1.1
python-decouple==3.8